Refactor RecordGenerator to utilize SchemaManager for schema entry management and improve code organization
